商品情報編集機能の実装

+++ app/controllers/items_controller.rb
@@ -23,13 +23,22 @@ class ItemsController < ApplicationController
   
 
   def destroy
-    item = Item.find(params[:id])
-    item.destroy
+    if item = Item.find(params[:id])
+       item.destroy
+       redirect_to root_path
+    else
+       render :edit
+    end
   end
 
   def update
-    item = Item.find(params[:id])
-    item.update(item_params)
+     if item = Item.find(params[:id])
+        item.update(item_params)
+        redirect_to root_path
+     else
+         render :edit
+     end
+    
   end
 
   def edit
